Add TopicSearch for finding paths through friends
This is the mother load. I've pulled out a reasonable amount of
the work just so that I could sanely work on this. But given that
time is of the essence I've still left it quite complex.

What I'd really like to see here is some dependency injection so
that I'm not limited to full integration specs.