Add expert search functionality to my profile page
This gives a simple search box that accepts a string which will
be used to query Passion topic. Any Members thus located will
be traced back to the current user's member profile through
the shortest connected list friends.

Results will be presented as a list of links to member profiles,
with the "expert" passion shown at the end.