Add AllMemberPreview query
This is how we want to show the list of our Members. We need their
id (for profile page link) their name, a website short url, and
a count of friends. Since the friendships could exist on either
side on the record we need to do some fancy counting to get the
sum.