Allow vertices along descending member IDs
This processing of the graph was not helpful. What it was actually
doing is ensuring that we could always travel "up" member IDs
through friendship. That's fine if you always happen go from a
small ID to a large ID (as I did when testing locally), but fails
to allow traversal down from a higher ID to a lower ID.

It's okay that this spec now shows "no match" because it's actually
the previous commit which is better connecting our friend graph
by duplicating all the edges to allow travel in either direction.
We need to remove this "bad fix" so that BOTH of those directions
will be searched by the gem.